fix ActiveTextureScope param, switched Texture::bind()/unbind() var type to uint8_t to hopefully emit a warning if value type is wrong

+++ include/cinder/gl/Texture.h
@@ -120,9 +120,9 @@ class Texture
 	//!	Disables the Texture's target
 	void			disable() const;
 	//!	Binds the Texture's texture to its target in the multitexturing unit \c GL_TEXTURE0 + \a textureUnit
-	void 			bind( GLuint textureUnit = 0 ) const;
+	void 			bind( uint8_t textureUnit = 0 ) const;
 	//!	Unbinds the Texture currently bound in the Texture's target
-	void			unbind( GLuint textureUnit = 0 ) const;
+	void			unbind( uint8_t textureUnit = 0 ) const;
 	
 	//!	Creates a new Texture from raw DirectDraw Stream data
 	static Texture	loadDds( IStreamRef ddsStream, Format format );
